pyproject.toml
[tool.poetry]
name = "erucakra"
version = "0.0.1"
description = "Climate Tipping Point Dynamics - A physically-motivated dynamical systems model for analyzing climate tipping points under various SSP scenarios"
authors = ["Sandy H. S. Herho <sandy.herho@email.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/sandyherho/erucakra"
repository = "https://github.com/sandyherho/erucakra"
keywords = ["climate", "tipping-points", "dynamical-systems", "SSP", "IPCC", "earth-system"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "Topic :: Scientific/Engineering :: Physics",
]
packages = [{include = "erucakra", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.0"
scipy = "^1.11.0"
pandas = "^2.0.0"
matplotlib = "^3.7.0"
netCDF4 = "^1.6.0"
xarray = "^*"
tqdm = "^4.65.0"
pyyaml = "^6.0"
click = "^8.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"
ruff = "^0.0.290"

[tool.poetry.scripts]
erucakra = "erucakra.cli:main"
erucakra-run = "erucakra.cli:run"
erucakra-list = "erucakra.cli:list_scenarios"
erucakra-info = "erucakra.cli:info"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "D", "UP", "B", "C4"]
ignore = ["D100", "D104"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
